Transfer lists to separate class
